#741: hexStringFromColor() 이동 및 수정
- 공용으로 사용할 수 있게 이동 및 수정하였습니다.
