+++ ToDo-Garden/ToDo-Garden/PostGroupScene/Sources/PostGroupScene/PostGroupWorker.swift
@@ -21,23 +21,3 @@ public struct PostGroupWorker: PostGroupWorkable {
     return group
   }
 }
-
-extension PostGroupWorker {
-  private func hexStringFromColor(_ color: UIColor) -> String {
-    let multiplier: CGFloat = 255.0
-    let redShift: Int = 16
-    let greenShift: Int = 8
-    
-    var red: CGFloat = CGFloat.zero
-    var green: CGFloat = CGFloat.zero
-    var blue: CGFloat = CGFloat.zero
-    var alpha: CGFloat = CGFloat.zero
-    
-    color.getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-    let rgb: Int = (Int(red * multiplier) << redShift) |
-    (Int(green * multiplier) << greenShift) |
-    (Int(blue * multiplier))
-    
-    return String(format: "#%06x", rgb)
-  }
-}
